+++ libs/world/src/system/mod.rs
@@ -12,7 +12,4 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with OpenFA.  If not, see <http://www.gnu.org/licenses/>.
-pub mod flight_dynamics;
 pub mod shape_mesh;
-pub mod transform;
-pub mod wheeled_dynamics;
